common/django_common/requirements.txt
Django>=5.2.5
djangorestframework>=3.16.1
django-filter>=25.1
psycopg2-binary>=2.9.10
python-dotenv>=1.1.1
django-redis>=5.4.0
redis>=6.4.0
gunicorn>=23.0.0
